Implemented exceptions in logic. Removed the updateModel for update simplicity purposes. Deleted isDeleted property with the corresponding methods.

+++ src/Logic/Logic.Core/TodoListLogic.cs
@@ -1,5 +1,6 @@
 ﻿namespace devdeer.ListOfWork.Logic.Core
 {
+    using devdeer.ListOfWork.Logic.Common.Exceptions;
     using devdeer.ListOfWork.Logic.Interfaces;
     using devdeer.ListOfWork.Logic.Models;
     using devdeer.ListOfWork.Repositories.Interfaces;
@@ -19,18 +20,28 @@
         /// <inheritdoc />
         public async Task<TodoItemModel> CreateTodoAsync(CreateTodoItemModel createTodo)
         {
+            if(string.IsNullOrEmpty(createTodo.Title) || string.IsNullOrWhiteSpace(createTodo.Title))
+            {
+                throw new ArgumentException("Value can't be null or whitespace.", nameof(createTodo.Title));
+            }
+            if (string.IsNullOrEmpty(createTodo.Description) || string.IsNullOrWhiteSpace(createTodo.Description))
+            {
+                throw new ArgumentException("Value can't be null or whitespace.", nameof(createTodo.Description));
+            }
+            if (createTodo.DueTime <= DateTimeOffset.Now)
+            {
+                throw new ArgumentException("Value can't have a value that is earlier or equals the time at the moment", nameof(createTodo.DueTime));
+            }
             return await Repository.CreateTodoAsync(createTodo);
         }
         /// <inheritdoc />
         public async Task<bool> DeleteTodoAsync(string id)
         {
+            if (string.IsNullOrEmpty(id) || string.IsNullOrWhiteSpace(id))
+            {
+                throw new ArgumentException("Value can't be null or whitespace.", nameof(id));
+            }
             return await Repository.DeleteTodoAsync(id);
-        }
-        /// <inheritdoc />
-        public async Task<IEnumerable<TodoItemModel>> GetAllDeletedTodosAsync()
-        {
-            var result = await Repository.GetAllTodosAsync();
-            return result!.Where(todo => todo.IsDeleted);
         }
         /// <inheritdoc />
         public async Task<IEnumerable<TodoItemModel>> GetAllTodosAsync()
@@ -40,16 +51,42 @@
         /// <inheritdoc />
         public async Task<TodoItemModel?> GetTodoByIdAsync(string id)
         {
+            if (string.IsNullOrEmpty(id) || string.IsNullOrWhiteSpace(id))
+            {
+                throw new ArgumentException("Value can't be null or whitespace.", nameof(id));
+            }
             return await Repository.GetByIdAsync(id);
         }
         /// <inheritdoc />
         public async Task<TodoItemModel?> SetTodoToCompleteAsync(string id)
         {
-            return await Repository.UpdateTodoAsync(new UpdateTodoItemModel());
+            if (string.IsNullOrEmpty(id) || string.IsNullOrWhiteSpace(id))
+            {
+                throw new ArgumentException("Value can't be null or whitespace.", nameof(id));
+            }
+            var todo = await GetTodoByIdAsync(id) ?? throw new EntityNotFoundException(id);
+            todo.CompletedAt = DateTimeOffset.UtcNow;
+            return await Repository.UpdateTodoAsync(todo);
         }
         /// <inheritdoc />
-        public async Task<TodoItemModel?> UpdateTodoAsync(UpdateTodoItemModel updateTodoItem)
+        public async Task<TodoItemModel?> UpdateTodoAsync(TodoItemModel updateTodoItem)
         {
+            if (string.IsNullOrEmpty(updateTodoItem.Id) || string.IsNullOrWhiteSpace(updateTodoItem.Id))
+            {
+                throw new ArgumentException("Value can't be null or whitespace.", nameof(updateTodoItem.Id));
+            }
+            if (string.IsNullOrEmpty(updateTodoItem.Title) || string.IsNullOrWhiteSpace(updateTodoItem.Title))
+            {
+                throw new ArgumentException("Value can't be null or whitespace.", nameof(updateTodoItem.Id));
+            }
+            if (string.IsNullOrEmpty(updateTodoItem.Description) || string.IsNullOrWhiteSpace(updateTodoItem.Description))
+            {
+                throw new ArgumentException("Value can't be null or whitespace.", nameof(updateTodoItem.Description));
+            }
+            if (updateTodoItem.DueTime <= DateTimeOffset.Now)
+            {
+                throw new ArgumentException("Value can't have a value that is earlier or equals the time at the moment", nameof(updateTodoItem.DueTime));
+            }
             return await Repository.UpdateTodoAsync(updateTodoItem);
         }
         /// <summary>
